moved eyex functionality into its own class, updated copyDLL scripts to run as part of the Visual Studio build, added *.VC.db to gitignore

+++ Code/TouchPoints/vs2015/CopyDebugx64Dlls.ps1
@@ -2,6 +2,6 @@
 
 if (!(Test-Path -path $debugFolder)) {New-Item $debugFolder -Type Directory}
 
-Copy-Item .\dependencies\LeapSDK\lib\x64\Leap.dll $debugFolder -Force -Verbose
-Copy-Item .\dependencies\libusb-1.0.20\x64\dll\libusb-1.0.dll $debugFolder -Force -Verbose
-Copy-Item .\dependencies\TobiiEyeXSDK_1.7.480\lib\x64\Tobii.EyeX.Client.dll $debugFolder -Force -Verbose
+if (!(Test-Path ($debugFolder + "Leap.dll") )) { Copy-Item .\dependencies\LeapSDK\lib\x64\Leap.dll $debugFolder -Force -Verbose }
+if (!(Test-Path ($debugFolder + "libusb-1.0.dll") )) { Copy-Item .\dependencies\libusb-1.0.20\x64\dll\libusb-1.0.dll $debugFolder -Force -Verbose }
+if (!(Test-Path ($debugFolder + "Tobii.EyeX.Client.dll") )) { Copy-Item .\dependencies\TobiiEyeXSDK_1.7.480\lib\x64\Tobii.EyeX.Client.dll $debugFolder -Force -Verbose }
